requirements.txt
requests==2.25.1
websocket-client=1.6.3
python-dotenv=1.0.0